Sets ctrlp's custom ignore not to ignore git directories
Ctrlp was failing to find files in a Go project, that was nested in a
directory named '.github.com'; a rather common pattern. The custom
ignore was matching on 'git' and ignoring the files. This regex is
borrowed from the ctrlp docs and updated with a couple of file and
directory types that I previously had excluded, and now finds files!