Ensure strict no remap; remove <ESC> remap
While trying to write recursive macros that would fail to recurse, I
noticed that the remap of <ESC> to include the :w command could cause
failures. I change this so <ESC> signals aren't swallowed. We strictly
avoid remap so that vim behaves as expected: see 'Learn Vim the Hard Way'